sdfasdf

+++ irs-maven/src/main/java/com/irs/pojo/TbBicycle.java
@@ -5,15 +5,15 @@ public class TbBicycle {
 
     private String name;
 
-    private Long number;
+    private Integer number;
 
-    private Long price;
+    private Integer price;
 
-    private Byte type;
+    private Integer type;
 
-    private Byte status;
+    private Integer status;
 
-    private Byte deflag;
+    private Integer deflag;
 
     public Integer getId() {
         return id;
@@ -31,43 +31,43 @@ public class TbBicycle {
         this.name = name == null ? null : name.trim();
     }
 
-    public Long getNumber() {
+    public Integer getNumber() {
         return number;
     }
 
-    public void setNumber(Long number) {
+    public void setNumber(Integer number) {
         this.number = number;
     }
 
-    public Long getPrice() {
+    public Integer getPrice() {
         return price;
     }
 
-    public void setPrice(Long price) {
+    public void setPrice(Integer price) {
         this.price = price;
     }
 
-    public Byte getType() {
+    public Integer getType() {
         return type;
     }
 
-    public void setType(Byte type) {
+    public void setType(Integer type) {
         this.type = type;
     }
 
-    public Byte getStatus() {
+    public Integer getStatus() {
         return status;
     }
 
-    public void setStatus(Byte status) {
+    public void setStatus(Integer status) {
         this.status = status;
     }
 
-    public Byte getDeflag() {
+    public Integer getDeflag() {
         return deflag;
     }
 
-    public void setDeflag(Byte deflag) {
+    public void setDeflag(Integer deflag) {
         this.deflag = deflag;
     }
 }
